added ability to add show and delete post from home page

+++ client/src/Components/reusuables/post/HomeFeedText.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { useAuthContext } from "../../../Hooks/useAuthContext";
 
 import { usePostsContext } from "../../../Hooks/usePostsContext";
 import ItemUserInfo from "../ItemUserInfo";
@@ -9,12 +10,17 @@ import PostLikeDisLike from "./PostLikeDislike";
 // Test
 import test from "../../../Assets/Images/user.jpeg";
 
-const HomeFeedText = ({ user, pic, postedDate, post, modal }) => {
+const HomeFeedText = ({ pic, postedDate, post, modal }) => {
   const { dispatch } = usePostsContext;
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch("/api/post/" + post._id, {
       method: "DELETE",
+      headers: { Authorization: `Bearer ${user.token}` },
     });
     const json = await response.json();
 
@@ -38,6 +44,8 @@ const HomeFeedText = ({ user, pic, postedDate, post, modal }) => {
       <h3 className="item-user-content">{post.title}</h3>
 
       <h5 className="item-user-status text-gray">{post.status}</h5>
+      <h1 onClick={handleClick}>delete</h1>
+      {/* This is the item that will delete the post FOR NOW */}
       <PostLikeDisLike />
     </section>
   );
